requirements.txt
langchain==0.1.0
langchain-community==0.0.28
langchain-openai==0.0.7
openai==1.3.5
faiss-cpu==1.7.4
pypdf==3.17.0
streamlit==1.30.0
streamlit-pdf-viewer==0.0.4
python-dotenv==1.0.1